+++ lib/mix/test/fixtures/configs/bad_import.exs
@@ -1,0 +1,3 @@
+import Mix.Config
+
+import_config "bad_root.exs"