+++ lib/mix/test/fixtures/configs/good_import.exs
@@ -1,0 +1,3 @@
+import Mix.Config
+import_config "good_config.exs"
+:done